requirements.txt
asgiref==3.7.2
cffi==1.16.0
Django==4.2.5
greenlet==2.0.2
pika==1.3.2
psycopg2-binary==2.9.8
pycparser==2.21
python-dotenv==1.0.0
rabbitmq==0.2.0
sqlparse==0.4.4
typing_extensions==4.8.0
